+++ apps/settings/templates/settings/personal/security/password.php
